requirements.txt
anthropic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
qdrant-client
openai
twilio
python-dotenv
langgraph
langchain-core
langchain-anthropic
